% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climateLocs2016}
\alias{climateLocs2016}
\title{Climate locations (December 2016)}
\format{An object of class \code{data.frame} with 8735 rows and 19 columns.}
\usage{
climateLocs2016
}
\description{
Climate locations for environment canada
}
\keyword{datasets}

