% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{climatelong}
\alias{climatelong}
\title{DEPRECATED: Transform EC data to long format}
\usage{
climatelong(df, rm.na = FALSE)
}
\arguments{
\item{df}{A wide data frame (obtained from \link{getClimateData} or \link{getClimateDataRaw})}

\item{rm.na}{Flag to remove rows with an empty value. This may help compress large datasets.}
}
\value{
A melted \code{data.frame} (see reshape2::melt)
}
\description{
DEPRECATED: Transform EC data to long format
}
