% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{ec_climate_data_read_csv}
\alias{ec_climate_data_read_csv}
\title{Read a historical climate data CSV file}
\usage{
ec_climate_data_read_csv(path)
}
\arguments{
\item{path}{A path to an (unmodified) CSV file downloaded from the EC service}
}
\value{
A tibble with an attribute "flag_info", containing the flag information.
}
\description{
Read a historical climate data CSV file
}
