% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getClimateDataRaw}
\alias{getClimateDataRaw}
\title{Get parsed CSV data from Environment Canada}
\usage{
getClimateDataRaw(stationID, timeframe = c("monthly", "daily", "hourly"),
  Year = NA, Month = NA,
  endpoint = "http://climate.weather.gc.ca/climate_data/bulk_data_e.html",
  ...)
}
\arguments{
\item{stationID}{A stationID (you could find this using \link{getsites})}

\item{timeframe}{One of "montly" "daily" or "hourly"}

\item{Year}{The year for which to fetch the data}

\item{Month}{The month for which to fetch the data}

\item{...}{further arguments passed on to the downloading function}

\item{Day}{The day for which to fetch the data}
}
\value{
A data.frame of results
}
\description{
This function just donloads a .csv file from the bulk data service from
Environment Canda. It follows as closely as possible the EC specifications,
and does not modify the result except to remove the header information. To
apply this function over multiple months/stations/years/months, use \link{getClimateData}.
}
\examples{
getClimateData(27141, timeframe="monthly")

}

