% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{getClimateSites}
\alias{getClimateSites}
\title{Get Environment Canada locations.}
\usage{
getClimateSites(location, year = NULL, n = 5, locs = NULL,
  nicenames = FALSE, cols = c("Name", "Province", "Station ID", "distance",
  "Latitude (Decimal Degrees)", "Longitude (Decimal Degrees)", "First Year",
  "Last Year"))
}
\arguments{
\item{location}{A human-readable location that will be geocoded}

\item{year}{A vector of years that the location must have data for}

\item{n}{The number of rows to return}

\item{locs}{The data.frame of locations. Use NULL to for \link{ecclimatelocs}.}

\item{nicenames}{Sanitize names to type-able form}

\item{cols}{The columns to return (use NULL for all columns)}
}
\value{
A subset of \link{ecclimatelocs}
}
\description{
Get Environment Canada locations.
}
\references{
\url{ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Readme.txt}
\url{ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/}
}
