% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Get Data from Environment Canada}
\usage{
getdata(stationID, timeframe = c("monthly", "daily", "hourly"), year = NULL,
  month = NULL, cache = NULL, quiet = TRUE, progress = c("text", "none",
  "tk"), format = c("wide", "long"))
}
\arguments{
\item{stationID}{Vector of stationIDs to get data from}

\item{timeframe}{Resolution of acquired data}

\item{year}{Vector of years (if applicable)}

\item{month}{Vector of months (if applicable)}
}
\value{
A data.frame of the results
}
\description{
Get Data from Environment Canada
}

