% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdataraw}
\alias{getdataraw}
\title{Get parsed CSV data from Environment Canada}
\usage{
getdataraw(stationID, timeframe = c("monthly", "daily", "hourly"),
  year = NA, month = NA)
}
\arguments{
\item{stationID}{A stationID (you could find this using \link{getsites})}

\item{timeframe}{One of "montly" "daily" or "hourly"}

\item{year}{The year for which to fetch the data}

\item{month}{The month for which to fetch the data}

\item{day}{The day for which to fetch the data}
}
\value{
A data.frame of results
}
\description{
Get parsed CSV data from Environment Canada
}

