% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsites.R
\name{getsites}
\alias{getsites}
\title{Get Environment canada data}
\usage{
getsites(location, years = NULL, n = 5, locs = climateLocs2016,
  cols = c("Name", "Province", "Station ID", "Latitude (Decimal Degrees)",
  "Longitude (Decimal Degrees)"))
}
\arguments{
\item{location}{A human-readable location that will be geocoded}

\item{years}{A vector of years that the location must have data for}

\item{n}{The number of rows to return}
}
\value{
A subset of \link{climateLocs2016}
}
\description{
Get Environment canada data
}
\examples{
getsites("Wolfville, NS")

}

