% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nice_names}
\alias{nice_names}
\alias{set_nice_names}
\title{Convert object names to programmer-friendly names}
\usage{
nice_names(x)

set_nice_names(x)
}
\arguments{
\item{x}{A vector of names (nice_names()) or an object with names (set_nice_names())}
}
\value{
A new vector of names or an object with the names attribute modified
}
\description{
Convert object names to programmer-friendly names
}
\examples{
nice_names(c("Name 1", "name2", "name_3", "NAME     4 "))

}
