% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_locations.R
\name{print.ec_climate_location_search}
\alias{print.ec_climate_location_search}
\title{Print climate location search results}
\usage{
\method{print}{ec_climate_location_search}(x, limit = 20, ...)
}
\arguments{
\item{x}{A climate location search result from \link{ec_climate_search_locations} or
\link{ec_climate_geosearch_locations}.}

\item{limit}{The number of results to show (use NULL for no limit)}

\item{...}{Not used.}
}
\value{
The input, invisibly.
}
\description{
Print climate location search results
}
