% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached results}
\usage{
clear_cache(cache = "ec.cache")
}
\description{
Clears the local cache of downloaded files
(located at \code{.api_result} in the \code{.GlobalEnv}).
}
\examples{
clear_cache()

}

