% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached results}
\usage{
clear_cache(cache = "ec.cache")
}
\arguments{
\item{cache}{A folder name where cached files should be stored.}
}
\description{
Clears the local cache of downloaded files (by default in the
ec.cache folder in the working directory).
}
\examples{
clear_cache()

}
