% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_getdata.R
\name{climatelong}
\alias{climatelong}
\title{Transform EC data to long format}
\usage{
climatelong(df, rm.na = FALSE)
}
\arguments{
\item{df}{A wide data frame (obtained from \link{getClimateData} or \link{getClimateDataRaw})}

\item{rm.na}{Flag to remove rows with an empty value. This may help compress large datasets.}
}
\value{
A melted \code{data.frame} (see reshape2::melt)
}
\description{
Transform EC data to long format
}
\examples{
# don't test because fetching of file slows down testing
\donttest{
df <- getClimateData(27141, timeframe="daily", year=2014:2016)
climatelong(df)
# also works with nicenames=TRUE
df <- getClimateData(27141, timeframe="daily", year=2014:2016, nicenames=TRUE)
climatelong(df)
}

}

