% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{ec_climate_data}
\alias{ec_climate_data}
\alias{ec_climate_data_base}
\title{Load Environment Canada Historical Climate Data}
\usage{
ec_climate_data(location, timeframe = c("monthly", "daily", "hourly"),
  year = NULL, month = NULL, day = NULL, cache = "ec.cache",
  quiet = TRUE)

ec_climate_data_base(location, timeframe = c("monthly", "daily", "hourly"),
  year = NULL, month = NULL, cache = NULL, quiet = FALSE,
  endpoint = "http://climate.weather.gc.ca/climate_data/bulk_data_e.html")
}
\arguments{
\item{location}{An unambiguous name identifier or station ID
(resolved using \link{as_ec_climate_location}).}

\item{timeframe}{One of monthly, daily, or hourly.}

\item{year}{A year or vector of years}

\item{month}{A month or vector of months}

\item{day}{A day or vector of days}

\item{cache}{A directory in which to cache downloaded files}

\item{quiet}{Use FALSE for verbose output}

\item{endpoint}{The web address for the EC data service}
}
\value{
A data.frame (tibble) with an attribute "flag_info", containing the flag information.
}
\description{
Load Environment Canada Historical Climate Data
}
