% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_locations.R
\name{ec_climate_search_locations}
\alias{ec_climate_search_locations}
\title{Search climate locations}
\usage{
ec_climate_search_locations(query = NULL, ..., timeframe = c("NA",
  "monthly", "daily", "hourly"), year = NULL, limit = NULL)
}
\arguments{
\item{query}{A query in several forms}

\item{...}{Additional arguments are used to \link[dplyr]{filter}
\link{ec_climate_locations_all}}

\item{timeframe}{The target timeframe for the query}

\item{year}{An optional year when the location must have data}

\item{limit}{The maximum number of locations to return (or NULL for no limit).
Lon/lat queries are automatically capped at 30 locations.}
}
\description{
Search climate locations
}
