% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{getClimateDataRaw}
\alias{getClimateDataRaw}
\title{DEPRECATED: Get parsed CSV data from Environment Canada}
\usage{
getClimateDataRaw(stationID, timeframe = c("monthly", "daily", "hourly"),
  Year = NA, Month = NA,
  endpoint = "http://climate.weather.gc.ca/climate_data/bulk_data_e.html",
  flag.info = FALSE, ...)
}
\arguments{
\item{stationID}{A stationID (you could find this using \link{getClimateSites})}

\item{timeframe}{One of "montly" "daily" or "hourly"}

\item{Year}{The year for which to fetch the data}

\item{Month}{The month for which to fetch the data}

\item{endpoint}{The url from which to fetch data (in case this changes in the future)}

\item{flag.info}{Pass TRUE to get a \code{list} with elements \code{$data} and \code{$flags}}

\item{...}{further arguments passed on to the downloading function}
}
\value{
A data.frame of results, or a list if flag.info=TRUE
}
\description{
This function just downloads a .csv file from the bulk data service from
Environment Canda. It follows as closely as possible the EC specifications,
and does not modify the result except to remove the header information. To
apply this function over multiple months/stations/years/months, use \link{getClimateData}.
}
\references{
\url{http://climate.weather.gc.ca/historical_data/search_historic_data_e.html}
\url{ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Readme.txt}
}
