% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsites.R
\name{getsites}
\alias{getsites}
\title{Get Environment canada data}
\usage{
getsites(location, year = NULL, n = 5, locs = NULL, cols = c("Name",
  "Province", "Station ID", "Latitude (Decimal Degrees)",
  "Longitude (Decimal Degrees)"))
}
\arguments{
\item{location}{A human-readable location that will be geocoded}

\item{year}{A vector of years that the location must have data for}

\item{n}{The number of rows to return}

\item{locs}{The data.frame of locations. Use NULL to for \link{climateLocs2016}.}
}
\value{
A subset of \link{climateLocs2016}
}
\description{
Get Environment canada data
}
\examples{
getsites("Wolfville, NS", year=2016)

}

